pyproject.toml
[tool.poetry]
name = "git init"
version = "0.1.0"
description = ""
authors = ["zebuth95"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"
redis = "^4.5.5"
slowapi = "^0.1.8"
sqlalchemy = "^2.0.15"
flake8 = "^6.0.0"
black = "^23.3.0"
python-dotenv = "^1.0.0"
bcrypt = "^4.0.1"
pyjwt = "^2.7.0"
pydantic = "^1.10.9"
email-validator = "^2.0.0.post2"
python-multipart = "^0.0.6"
pydantic-computed = "^0.2.2"
py-fastapi-logging = "^0.3.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
